IFEDMethodBase, part 2.
