+++ source/interaction/ifed_method_base.cc
@@ -1,9 +1,15 @@
 #include <fiddle/interaction/ifed_method_base.h>
 
+#include <limits>
+
 namespace fdl
 {
-  using namespace dealii;
-  using namespace SAMRAI;
+  template <int dim, int spacedim>
+  IFEDMethodBase<dim, spacedim>::IFEDMethodBase()
+    : current_time(std::numeric_limits<double>::signaling_NaN())
+    , half_time(std::numeric_limits<double>::signaling_NaN())
+    , new_time(std::numeric_limits<double>::signaling_NaN())
+  {}
 
   template class IFEDMethodBase<NDIM, NDIM>;
 } // namespace fdl
